New concept "editable / enabled" and remove xxxNotEditable functions
This change completely removes the "clickable" concept from AludraTest.
Especially in web, everything is clickable. The implicit checks for 
disabled or other states now must be performed using assertXxx or
isXxx methods, which have been enriched.

The "editable" state now only applies for text-based input components,
which do NOT have a read-only flag set. 

For all components, a consistent "enabled" state marks that the 
component is NOT disabled within the GUI, i.e. accepts user interaction
and / or does not have a depiction marking it as diabled (e.g. grayed).